modificato auth in modo che salvi in locale i dati dell'utente attuale

+++ src/app/services/local-write.service.ts
@@ -1,24 +1,29 @@
-import { Injectable } from '@angular/core';
-
+import { Inject,Injectable, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 @Injectable({
   providedIn: 'root'
 })
 export class LocalWriteService {
-
-  constructor() { }
+  private isBrowser: boolean;
+constructor(@Inject(PLATFORM_ID) private platformId: Object) {
+    this.isBrowser = isPlatformBrowser(platformId);
+  }
   /* Saves a value in local storage under the specified key.
    * @param key - A string representing the key to associate with the value
    * @param value - The value to be stored in local storage.
   */
   public saveData(key:string,value:string):void{
-    localStorage.setItem(key,value);
+    if(this.isBrowser)
+      localStorage.setItem(key,value);
   }
   /* Retrieves the stored value present in  local storage for the specified key
   *@param key - The key to search for in local storage.
   *@returns The stored value if found; otherwise, null.
    */
   public getData(key:string):string | null{
-    return localStorage.getItem(key);
+    if(this.isBrowser)
+      return localStorage.getItem(key);
+    return null;
   }
 
 /**
@@ -26,11 +31,13 @@ export class LocalWriteService {
  * @param key - The key of the item to remove from local storage.
  */
   public removeData(key:string):void{
-    localStorage.removeItem(key);
+    if(this.isBrowser)
+      localStorage.removeItem(key);
   }
   /* Wipe out all the data in the local storage
    * */
   public clearData():void{
-    localStorage.clear();
+    if(this.isBrowser)
+      localStorage.clear();
   }
 }
